+++ src/main/java/com/example/homework20/repositories/Employees.java
@@ -1,0 +1,11 @@
+package com.example.homework20.repositories;
+
+import com.example.homework20.model.Employee;
+
+import java.util.HashMap;
+import java.util.Map;
+
+public class Employees {
+    public static final int MAX = 15;
+    public static final Map<String, Employee> employees = new HashMap<>();
+}
