+++ src/main/java/com/example/homework20/service/DepartmentService.java
@@ -1,0 +1,8 @@
+package com.example.homework20.service;
+
+public interface DepartmentService {
+    String getEmployeeWithMaxSalary(int departmentId);
+    String getEmployeeWithMinSalary(int departmentId);
+    String printAllEmployees(int departmentId);
+    String printAllEmployeesByDepartments();
+}
