+++ src/main/java/com/example/homework20/service/EmployeeServiceImpl.java
@@ -5,13 +5,11 @@ import com.example.homework20.exceptions.EmployeeNotFoundException;
 import com.example.homework20.exceptions.EmployeeStorageIsFullException;
 import com.example.homework20.model.Employee;
 
-import java.util.*;
-import java.util.stream.Collectors;
+import static com.example.homework20.repositories.Employees.MAX;
+import static com.example.homework20.repositories.Employees.employees;
 
 @org.springframework.stereotype.Service
 public class EmployeeServiceImpl implements EmployeeService {
-    private static final int MAX = 15;
-    private final Map<String, Employee> employees = new HashMap<>();
 
     @Override
     public void populate() {
@@ -66,50 +64,8 @@ public class EmployeeServiceImpl implements EmployeeService {
         return builder.toString();
     }
 
-    @Override
-    public String getEmployeeWithMinSalary(int department) {
-        return employees.values().stream()
-                .filter(e -> e.getDepartment() == department)
-                .min(Comparator.comparingInt(Employee::getSalary))
-                .orElseThrow(EmployeeNotFoundException::new)
-                .toString();
-    }
-
-    @Override
-    public String getEmployeeWithMaxSalary(int department) {
-        return employees.values().stream()
-                .filter(e -> e.getDepartment() == department)
-                .max(Comparator.comparingInt(Employee::getSalary))
-                .orElseThrow(EmployeeNotFoundException::new)
-                .toString();
-    }
-
-    @Override
-    public String printAllEmployees(int department) {
-        StringBuilder builder = new StringBuilder("<b>Сотрудники " + department + " отдела:</b>");
-        employees.values().stream()
-                .filter(e -> e.getDepartment() == department)
-                .forEach(e -> builder.append("<br>").append(e.toStringWithoutDepartment()));
-        return builder.toString();
-    }
-
-    @Override
-    public String printAllEmployeesByDepartments() {
-        StringBuilder builder = new StringBuilder();
-        getAllDepartments().forEach(d -> builder.append(printAllEmployees(d)).append("<br>"));
-        return builder.toString();
-    }
-
     public Employee getEmployee(String firstName, String lastName) {
         return employees.get(getKey(firstName, lastName));
-    }
-
-    public List<Integer> getAllDepartments() {
-        return employees.values().stream()
-                .map(Employee::getDepartment)
-                .distinct()
-                .sorted()
-                .collect(Collectors.toList());
     }
 
     private String getKey(String firstName, String lastName) {
